+++ src/main/java/basemod/BaseMod.java
@@ -10,19 +10,15 @@ import java.lang.reflect.Modifier;
 import java.lang.reflect.Type;
 import java.net.URISyntaxException;
 import java.net.URL;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
 import java.util.Map.Entry;
-import java.util.Properties;
-import java.util.Set;
-import java.util.function.Consumer;
+import java.util.stream.Collectors;
 
 import basemod.interfaces.*;
 import basemod.patches.whatmod.WhatMod;
 import com.evacipated.cardcrawl.modthespire.lib.SpireConfig;
 import com.megacrit.cardcrawl.helpers.*;
+import com.megacrit.cardcrawl.monsters.MonsterGroup;
 import com.megacrit.cardcrawl.rooms.MonsterRoom;
 import com.megacrit.cardcrawl.screens.custom.CustomModeCharacterButton;
 import org.apache.logging.log4j.LogManager;
@@ -90,7 +86,6 @@ import com.megacrit.cardcrawl.relics.Circlet;
 import com.megacrit.cardcrawl.rooms.AbstractRoom;
 import com.megacrit.cardcrawl.saveAndContinue.SaveAndContinue;
 import com.megacrit.cardcrawl.screens.charSelect.CharacterOption;
-import com.megacrit.cardcrawl.screens.mainMenu.MainMenuScreen;
 import com.megacrit.cardcrawl.screens.stats.CharStat;
 import com.megacrit.cardcrawl.shop.ShopScreen;
 import com.megacrit.cardcrawl.shop.StorePotion;
@@ -1174,7 +1169,99 @@ public class BaseMod {
 			return null;
 		}
 	}
-	
+
+	//
+	// Monsters
+	//
+
+	private static HashMap<String, GetMonsterGroup> customMonsterEncounters = new HashMap<>();
+
+	public interface GetMonsterGroup {
+		MonsterGroup get();
+	}
+
+	public interface GetMonster {
+		AbstractMonster get();
+	}
+
+	public static void addMonsterEncounter(String encounterID, GetMonster monster) {
+		customMonsterEncounters.put(encounterID, () -> new MonsterGroup(monster.get()));
+	}
+
+	public static void addMonsterEncounter(String encounterID, GetMonsterGroup group) {
+		customMonsterEncounters.put(encounterID, group);
+	}
+
+	public static MonsterGroup getMonsterEncounter(String encounterID) {
+		GetMonsterGroup getter = customMonsterEncounters.get(encounterID);
+		if (getter == null) {
+			return null;
+		}
+		return getter.get();
+	}
+
+	public static boolean customMonsterEncounterExists(String encounterID) {
+		return customMonsterEncounters.containsKey(encounterID);
+	}
+
+	//
+	// Bosses
+	//
+
+	private static HashMap<String, List<BossInfo>> customBosses = new HashMap<>();
+
+	public static class BossInfo {
+		public final String id;
+		public final Texture bossMap;
+		public final Texture bossMapOutline;
+
+		private BossInfo(String id, String mapIcon, String mapIconOutline) {
+			this.id = id;
+			if (mapIcon != null) {
+				bossMap = ImageMaster.loadImage(mapIcon);
+			} else {
+				bossMap = null;
+			}
+			if (mapIconOutline != null) {
+				bossMapOutline = ImageMaster.loadImage(mapIconOutline);
+			} else {
+				bossMapOutline = null;
+			}
+		}
+	}
+
+	public static void addBoss(String dungeon, String bossID, String mapIcon, String mapIconOutline) {
+		if (!customBosses.containsKey(dungeon)) {
+			customBosses.put(dungeon, new ArrayList<>());
+		}
+		BossInfo info = new BossInfo(bossID, mapIcon, mapIconOutline);
+		customBosses.get(dungeon).add(info);
+	}
+
+	public static List<String> getBossIDs(String dungeonID) {
+		if (customBosses.containsKey(dungeonID)) {
+			return customBosses.get(dungeonID).stream()
+					.map(info -> info.id)
+					.collect(Collectors.toList());
+		}
+		return new ArrayList<>();
+	}
+
+	public static BossInfo getBossInfo(String bossID) {
+		if (bossID == null) {
+			return null;
+		}
+
+		for (List<BossInfo> infos : customBosses.values()) {
+			for (BossInfo info : infos) {
+				if (bossID.equals(info.id)) {
+					return info;
+				}
+			}
+		}
+		return null;
+	}
+
 	//
 	// Keywords
 	//